Increase defaults for `--pool-size` and `--max-simultaneous-connections-per-ip`

+++ src/entrypoint.rs
@@ -660,7 +660,7 @@ pub async fn entrypoint(config: ApplicationConfig) -> color_eyre::Result<()> {
         disable_tcp: config.disable_tcp,
         disable_aliasing: config.disable_aliasing,
         buffer_size,
-        pool_size: usize::from(config.pool_size),
+        pool_size: config.pool_size,
         max_connections_per_ip: usize::from(config.max_simultaneous_connections_per_ip),
         pool_timeout: config.pool_timeout,
         rate_limit: config
